move game.mjs and game.css into components folder
